requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
groq>=0.4.0
pdfplumber>=0.10.0
